test(cli): Refactor test helpers into modules
