+++ packages/kasten-cli/src/__test__/temp_dir.ts
@@ -1,0 +1,7 @@
+import { tmpdir } from "os";
+import { join } from "path";
+import * as fs from "fs";
+
+export function createTempDir(): string {
+  return fs.mkdtempSync(join(tmpdir(), "test-"));
+}
