+++ packages/kasten-cli/src/__test__/test_data.ts
@@ -1,0 +1,12 @@
+export const testNotes = [
+  {
+    relativePath: "note-1.mdx",
+    txt: "---\ntitle: Title 1\n---\nContent 1\n",
+    title: "Title 1",
+  },
+  {
+    relativePath: "note-2.mdx",
+    txt: "---\ntitle: Title 2\n---\nContent 2\n",
+    title: "Title 2",
+  },
+];
